requirements.txt
kafka-python
pandas
requests
sqlalchemy